% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOmegaCorr.R
\name{addOmegaCorr}
\alias{addOmegaCorr}
\title{add Omega correlations to a parameter table}
\usage{
addOmegaCorr(pars, by = NULL, as.fun)
}
\arguments{
\item{pars}{A parameter table, like returned by `NMreadExt()`.}

\item{by}{The name of a column, as a string. Calculate the
correlations within a grouping variable?  This will often be a
column containing the model name.}

\item{as.fun}{See `?NMdataConf`}
}
\value{
The parameter table with a `corr` column added.
}
\description{
add Omega correlations to a parameter table
}
