% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParType.R
\name{addParameter}
\alias{addParameter}
\title{add parameter based on par.type and i,j}
\usage{
addParameter(pars, overwrite = FALSE)
}
\arguments{
\item{pars}{Table of parameters to augment with additional columns.}

\item{overwrite}{Overwrite non-missing values? Default is `FALSE`.}
}
\description{
Columns filled or overwritten: parameter, par.name.
}
\keyword{internal}
