% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTAPD.R
\name{addTAPD}
\alias{addTAPD}
\title{Add cumulative number of doses, time of last dose and time since last dose to data}
\usage{
addTAPD(
  data,
  col.time = "TIME",
  col.tpdos = "TPDOS",
  col.tapd = "TAPD",
  col.ndoses = "NDOSES",
  col.evid = "EVID",
  col.amt = "AMT",
  col.pdosamt = "PDOSAMT",
  col.doscuma = "DOSCUMA",
  subset.dos,
  subset.is.complete,
  order.evid = c(3, 0, 2, 4, 1),
  by = "ID",
  as.fun
)
}
\arguments{
\item{data}{The data set to add the variables to.}

\item{col.time}{Name of time column (created by addTAPD). Default
it TIME.}

\item{col.tpdos}{Name of the time of previous dose column (created
by addTAPD). Default is TPDOS. Set to NULL to not create this
column.}

\item{col.tapd}{Name of the time of prvious dose column (created
by addTAPD). Default is TAPD. Set to NULL to not create this
column.}

\item{col.ndoses}{The name of the column (created by addTAPD) that
holds the cumulative number of doses administered to the
subject. Set to NULL to not create this column.}

\item{col.evid}{The name of the event ID column. This must exist
in data. Default is EVID.}

\item{col.amt}{col.evid The name of the dose amount column. This
must exist in data. Default is AMT.}

\item{col.pdosamt}{The name of the column to be created holding
the previous dose amount. Set to NULL to not create this
column.}

\item{col.doscuma}{The name of the column to be created holding
the cumulative dose amount. Set to NULL to not create this
column.}

\item{subset.dos}{A string that will be evaluated as a custom
expression to identify relevant events. See subset.is.complete
as well.}

\item{subset.is.complete}{Only used in combination with
non-missing subset.dos. By default, subset.dos is used in
addition to the impact of col.evid and col.amt. If
subset.is.complete=TRUE, subset.dos is used alone, and
col.evid and col.amt are completely ignored. This is typically
useful if the events are not doses but other events that are
not expressed as a typical dose combination of EVID and AMT
columns.}

\item{order.evid}{Order of events. This will only matter if there
are simultaneous events of different event types within
subjects. Typically if using nominal time, it may be important
to specify whether samples at dosing times are pre-dose
samples. The default is c(3,0,4,1,2) - i.e. samples and
simulations are pre-dose.}

\item{by}{Columns to do calculations within. Default is ID.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say tibble::as_tibble) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}
}
\description{
For now, doses have to be in data as EVID=1 and/or EVID=4
records. They can be in the format of one row per dose or repeated
dosing notation using ADDL and II.
}
\details{
This is still experimental.
}
