% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSA.R
\name{BSA}
\alias{BSA}
\title{Calculate BSA with different methods}
\usage{
BSA(weight, height, method)
}
\arguments{
\item{weight}{weight in KG}

\item{height}{height in CM}

\item{method}{method to calculate BSA}
}
\description{
Calculate BSA with different methods
}
\keyword{BSA}
