% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColRow.R
\name{checkColRow}
\alias{checkColRow}
\title{check that col.row is not edited in Nonmem control stream}
\usage{
checkColRow(col.row, file)
}
\arguments{
\item{col.row}{The name of the unique row identifier (say "ROW").}

\item{file}{a list file or input control stream file path.}
}
\value{
TRUE if no issues found
}
\description{
In order to safely merge by a unique row identifier,
    that row identifier must not be edited from input to
    output. checkColRow helps checking that based on the control
    stream.
}
\keyword{internal}
