% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{cl}
\alias{cl}
\title{Define a vector with factor levels in the same order as occurring in the vector.}
\usage{
cl(...)
}
\arguments{
\item{...}{unique elements or vectors with unique elements}
}
\description{
This is a shortcut for creating factors with levels as the order
of appearance of the specified levels.
}
\examples{
factor("b","a")
cl("b","a")
x <- c("b","a")
factor(x)
cl(x)
}
\seealso{
cc
}
