% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFileNames.R
\name{cleanFileNames}
\alias{cleanFileNames}
\title{Drop spaces and odd characters. Use to ensure generated file names are
usable.}
\usage{
cleanFileNames(x, allow.slash = FALSE)
}
\arguments{
\item{x}{a string to clean}

\item{allow.slash}{Allow slashes? Probably set to TRUE if x can be a
directory.}
}
\value{
A character vector
}
\description{
Drop spaces and odd characters. Use to ensure generated file names are
usable.
}
\examples{
cleanFileNames("e w\% # ff!l3:t,3?.csv")
cleanFileNames("3!?:#;<>=, {}|=g+&-
.csv")
}
\keyword{internal}
