% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSpaces.R
\name{cleanSpaces}
\alias{cleanSpaces}
\title{Drop leading, trailing and repeated spaces in character strings}
\usage{
cleanSpaces(x, double = TRUE, lead = TRUE, trail = TRUE)
}
\arguments{
\item{x}{A vector of character strings to modify}

\item{double}{Replace any number of consecutive blank spaces by a
single blank. Default is TRUE.}

\item{lead}{Drop spaces before first non-empty character. Default
is TRUE.}

\item{trail}{Drop spaces after last non-empty character. Default
is TRUE.}
}
\description{
Drop leading, trailing and repeated spaces in character strings
}
\keyword{internal}
