% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsenv.R
\name{clsenv}
\alias{clsenv}
\title{Clearing the global environment}
\usage{
clsenv(keep.vals, rm = T)
}
\arguments{
\item{keep.vals}{Vector of values that should be kept in the globalenvironment}

\item{rm}{If not TRUE than function returns the values that would have been removed}
}
\description{
Clearing the global environment
}
