% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colLabels.R
\name{colLabels}
\alias{colLabels}
\title{Extract column labels as defined in SAS}
\usage{
colLabels(x, sort = "alpha", as.fun)
}
\arguments{
\item{x}{object with elements containing label attributes.}

\item{sort}{If sort="alpha", results are sorted alphabetically.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function in as.fun to convert to something else. If
data.tables are wanted, use `as.fun="data.table"`. The default
can be configured using `NMdataConf()`.}
}
\value{
A data.frame with variable and their labels
}
\description{
Extract column labels as defined in SAS
}
\seealso{
compareCols NMinfo
}
