% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedArg.R
\name{deprecatedArg}
\alias{deprecatedArg}
\title{Report if an argument is deprecated.}
\usage{
deprecatedArg(oldarg, newarg, args, which = 2)
}
\arguments{
\item{oldarg}{The deprecated argument name (a character string).}

\item{newarg}{The non-deprecated argument name (a character string).}
}
\value{
The coalesced value of arguments
}
\description{
Only supposed to be called from within a function. For now only
works for arguments that have been replaced by others.
}
\keyword{internal}
