% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2mat.R
\name{dt2mat}
\alias{dt2mat}
\title{Convert a data.table of parameter estimates to a matrix}
\usage{
dt2mat(pars)
}
\arguments{
\item{pars}{A data.table with parameters like a subjset of the
parameter list returned by NMreadExt. It MUST ONLY INCLUDE THE
UNIQUE VALUES, so a upper or lower triangle of the matrix, not
the full matrix.}
}
\value{
a matrix
}
\description{
Often needed when using estimates of Omega or Sigma matrices in
further calculations.
}
