% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtapply.R
\name{dtapply}
\alias{dtapply}
\title{Apply function and return a data.table}
\usage{
dtapply(
  X,
  FUN,
  value.names = NULL,
  element.name = "element",
  fill = TRUE,
  as.fun,
  ...
)
}
\arguments{
\item{value.names}{If supplied, setnames will be run on each element returned by lapply wit value.names as the `new` argument.}

\item{...}{arguments passed to lapply}
}
\value{
a data.table
}
\description{
A convenience function that returns a data.table with a column##' representing the input values and a column with results. This is
still experimental and will not work for many input structures.
}
\details{
Only functions that return vectors are currently
    supported. dtapply should support functions that return
    data.frames.
}
\keyword{internal}
