% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtFillObs.R
\name{dtFillCols}
\alias{dtFillCols}
\title{Create a data table with columns of unequal length}
\usage{
dtFillCols(...)
}
\arguments{
\item{...}{Vectors to put into data.table}
}
\description{
Create a data table with columns of unequal length
}
\keyword{internal}
