% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadParsText.R
\name{escape.charclass}
\alias{escape.charclass}
\title{Escape selected special characters with backslashes}
\usage{
escape.charclass(x)
}
\arguments{
\item{x}{a character string}
}
\description{
Escape selected special characters with backslashes
}
\keyword{internal}
