% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sebas_readlst.R
\name{extract}
\alias{extract}
\title{Extract items from .lst and .ext files}
\usage{
extract(run)
}
\arguments{
\item{run}{Run number. Make sure you are in the right directory /(setwd/(/)/)}
}
\value{
OFV, cond_nr, run_date,final_est,term_message, etabar
}
\description{
Extract items from .lst and .ext files
}
