% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFM.R
\name{FFM}
\alias{FFM}
\title{Caculate the FFM}
\usage{
FFM(wt, height, gender, age = NULL)
}
\arguments{
\item{wt}{numeric}

\item{height}{numeric}

\item{gender}{1 for male and 0 for female, or "M/m" for male and "F/f" for female}

\item{age}{numeric}
}
\description{
Caculate the FFM
}
\details{
Functions from https://academic.oup.com/jac/article/73/8/2104/4995076
}
