% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgs.R
\name{getArgs}
\alias{getArgs}
\title{Get provided arguments as a named list}
\usage{
getArgs(which = 1)
}
\arguments{
\item{which}{How many environment levels to go up to look for
arguments.}
}
\value{
A named list
}
\description{
Get provided arguments as a named list
}
\seealso{
Other arguments: 
\code{\link{deprecatedArg}()}
}
\concept{arguments}
\keyword{internal}
