% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdata_utils.R
\name{getFileMod}
\alias{getFileMod}
\title{Determine file.mod to use based on argument and options}
\usage{
getFileMod(file.lst, file.mod = NULL)
}
\arguments{
\item{file.lst}{Path to output control stream.}

\item{file.mod}{Path or function. Default is NULL. See details.}
}
\value{
path to .mod file
}
\description{
Determine file.mod to use based on argument and options
}
\details{
NMdata needs the input control stream to find the path to
    the input data file. You have a few different options.
    \itemize{
\item{PSN style}{By default, NMdata assumes that
    by stripping the extension from the output control stream and
    appending .mod, it will find the input control stream.}
\item{path}{file.mod="path/to/input/control/strem"}
\item{translation function}{file.path can be a function that takes the lst path as argument and returns the input control stream path.}
\item{set option}{If you use a function, you may want to set this as default behaviour. Say your output control stream is always called input.txt and located in the same dir as the output control stream, you can use
options(NMdata.file.mod=function(file) file.path(dirname(file),"input.txt"))
}
}

Notice, if the argument dir.data is used in NMscanData or
NMtransInput, the input control stream is not used at all.
}
