% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilePaths.R
\name{getFilePaths}
\alias{getFilePaths}
\title{Internal interpretation of file specification options}
\usage{
getFilePaths(files = NULL, file.pattern = NULL, dir = NULL, quiet = FALSE)
}
\arguments{
\item{files}{character vector of full file paths. Specify either
files or both file.pattern and dir.}

\item{file.pattern}{A regular expression to look for in dir. If used, dir must also be supplied.}

\item{dir}{The directory i which to look for file.pattern. dir is
passed to list.files as pattern. If supplied, file.pattern must also be supplied.}
}
\value{
A character vector of full paths to files
}
\description{
Internal interpretation of file specification options
}
\keyword{internal}
