% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLines.R
\name{getLines}
\alias{getLines}
\title{read lines as needed}
\usage{
getLines(
  file,
  lines,
  linesep = "\\n",
  simplify = TRUE,
  ensure.ascii = TRUE,
  col.model,
  modelname,
  as.one
)
}
\arguments{
\item{file}{A file path to a a text file to read.}

\item{lines}{Text lines if file was already read.}

\item{ensure.ascii}{If `TRUE`, the result is converted to ascii characters.
This is done to ensure some further processing will work but you may
lose say greek characters from comments.}

\item{as.one}{If the `file` argument is used and if `as.one` is TRUE, the
file(s) are read and put into a `data.table` with a model column and a
`text` column. Default is FALSE. Be careful with this, as it returns
different formats whether the `file` or the `lines` argument is used.}
}
\description{
Functions that take file and lines arguments can use this function to derive
lines no matter what was provided.
}
\keyword{internal}
