% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadInits.R
\name{initsToExt}
\alias{initsToExt}
\title{Convert inits elements to a parameter data.frame}
\usage{
initsToExt(elements)
}
\arguments{
\item{elements}{The elements object produced by `NMreadInits()`.}
}
\description{
Convert inits elements to a parameter data.frame
}
\details{
initsToExt is misleading. It is not a reference to the
    initstab, but actually the elements object returned by
    NMreadInits. The elements object is more detailed as it
    contains information about where information is found in
    control stream lines. The `ext` object is a parameter
    `data.frame`, same format as returned by
    `NMdata::NMreadExt()`.
}
\keyword{internal}
