% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdata_utils.R
\name{is.NMdata}
\alias{is.NMdata}
\title{Check if an object is 'NMdata'}
\usage{
is.NMdata(x)
}
\arguments{
\item{x}{Any object}
}
\value{
logical if x is an 'NMdata' object
}
\description{
Check if an object is 'NMdata'
}
