% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSAME.R
\name{isSAME}
\alias{isSAME}
\title{Test whether an initial value is of type SAME}
\usage{
isSAME(x)
}
\arguments{
\item{x}{A character string to test}
}
\description{
Test whether an initial value is of type SAME
}
\examples{
isSame("SAME")
isSame(" SAME ")
isSame("SAME(1)")
isSame("SAME (1)")
isSame("  SAME   (  123 ) ")
isSame("SAME()")
isSame("SAME (a)")
isSame("NOT SAME")
}
\keyword{internal}
