% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMissings.R
\name{listMissings}
\alias{listMissings}
\title{List rows with missing values across multiple columns}
\usage{
listMissings(data, cols, by, na.strings = c("", "."), quiet = FALSE, as.fun)
}
\arguments{
\item{data}{The data to look into.}

\item{cols}{The columns to look for missings in.}

\item{by}{If supplied, we are keeping track of the missings within
the values of the by columns. In summary, by is included too.}

\item{na.strings}{Strings that should be interpreted as
missing. All spaces will be removed before we compare to
na.strings. The default is c("",".") so say " .  " is a
missing by default.}

\item{quiet}{Keep quiet? Default is not to.}

\item{as.fun}{A function that will be run un the result before
returning. If first input data set is a data.table, the
default is to return a data.table, if not the default is to
return a data.frame. Use whatever to get what fits in with
your workflow. Default can be configured with NMdataConf.}
}
\description{
Missing can be NA and for character variables it can be certain
strings too. This function is experimental and design may change
in future releases.
}
\seealso{
Other DataWrangling: 
\code{\link{compareCols}()},
\code{\link{dims}()}
}
\concept{DataWrangling}
