% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{load.data}
\alias{load.data}
\title{function to import data file based on file type (either 'xlsx' or 'csv')}
\usage{
load.data(name, file_type, na_value)
}
\arguments{
\item{name}{file name without extension}

\item{file_type}{Type of file xlsx or csv}
}
\description{
function to import data file based on file type (either 'xlsx' or 'csv')
}
\examples{
-
}
\keyword{csv}
\keyword{xlsx}
