% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{loaddata}
\alias{loaddata}
\title{function to import data file based on file type (either 'xlsx' or 'csv')}
\usage{
loaddata(name, na_value)
}
\arguments{
\item{name}{file name (can use tab to open, including file format)}

\item{na_value}{optional: argument to give the value that specifies NA}
}
\description{
function to import data file based on file type (either 'xlsx' or 'csv')
}
\examples{
-
}
\keyword{csv}
\keyword{xlsx}
