% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameMultipleFiles.R
\name{nameMultipleFiles}
\alias{nameMultipleFiles}
\title{Create file names for multiple list elements}
\usage{
nameMultipleFiles(fn, list.obj)
}
\arguments{
\item{fn}{File name to provide stem for all file names}

\item{list.obj}{List of objects to provide names for}
}
\description{
Create file names for multiple list elements
}
\keyword{internal}
