% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcheckColnames.R
\name{NMcheckColnames}
\alias{NMcheckColnames}
\title{Compare $INPUT in control stream to column names in input data}
\usage{
NMcheckColnames(file, as.fun, ...)
}
\arguments{
\item{file}{A Nonmem control stream or list file}

\item{as.fun}{See ?NMdataConf}

\item{...}{Additional arguments passed to}
}
\value{
An overview of input column names and how they are translated
}
\description{
Misspecification of column names in $DATA are a common source of
problems with Nonmem models, and one of the first things to check
when seemingly inexplainable things happen. This function lines up
input data column names with $DATA and how NMscanData will
interpret $DATA so you can easily spot if something is off.
}
\concept{debug}
