% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdata_utils.R
\name{NMdata_filepath}
\alias{NMdata_filepath}
\title{get a file or dir from a the NMdata package}
\usage{
NMdata_filepath(...)
}
\arguments{
\item{...}{parameters to pass to system.file}
}
\description{
get a file or dir from a the NMdata package
}
\details{
a light wrapper around system.file
}
