% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdataConf.R
\name{NMdataConfOptions}
\alias{NMdataConfOptions}
\title{Get NMdataConf parameter properties}
\usage{
NMdataConfOptions(name)
}
\arguments{
\item{name}{Optionally, a single parameter name (say "as.fun").}
}
\value{
If name is provided, a list representing one argument,
    otherwise a list with an element for each argument that can be
    customized using NMdataConf.
}
\description{
Get NMdataConf parameter properties
}
