% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdataConf.R
\name{NMdataDecideOption}
\alias{NMdataDecideOption}
\title{Determine active parameter value based on argument and NMdataConf setting}
\usage{
NMdataDecideOption(name, argument, allow.unknown = FALSE)
}
\arguments{
\item{name}{The name of the parameter, say "as.fun"}

\item{argument}{The value to pass. If missing or NULL, the value returned by NMdataConf/NMdataGetOption will typically be used.}
}
\value{
Active argument value.
}
\description{
Determine active parameter value based on argument and NMdataConf setting
}
\keyword{internal}
