% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdataConf.R
\name{NMdataGetOption}
\alias{NMdataGetOption}
\title{Look up default configuration of an argument}
\usage{
NMdataGetOption(...)
}
\arguments{
\item{...}{argument to look up. Only one argument can be looked up.}
}
\value{
The value active in configuration
}
\description{
Look up default configuration of an argument
}
