% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdata_utils.R
\name{NMdataOperations}
\alias{NMdataOperations}
\alias{merge.NMdata}
\alias{t.NMdata}
\alias{dimnames.NMdata}
\alias{rbind.NMdata}
\alias{cbind.NMdata}
\title{Basic arithmetic}
\usage{
\method{merge}{NMdata}(x, ...)

\method{t}{NMdata}(x, ...)

\method{dimnames}{NMdata}(x, ...)

\method{rbind}{NMdata}(x, ...)

\method{cbind}{NMdata}(x, ...)
}
\arguments{
\item{x}{an NMdata object}

\item{...}{arguments passed to other methods.}
}
\value{
An object that is not of class 'NMdata'.
}
\description{
Basic arithmetic
}
\details{
When 'dimnames', 'merge', 'cbind', 'rbind', or 't' is
    called on an 'NMdata' object, the 'NMdata' class is dropped,
    and then the operation is performed. So if and 'NMdata' object
    inherits from 'data.frame' and no other clases (which is
    default), these operations will be performed using the
    'data.frame' methods. But for example, if you use 'as.fun' to
    get a 'data.table' or 'tbl', their respective methods are used
    instead.
}
