% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMextractDataFile.R
\name{NMextractDataFile}
\alias{NMextractDataFile}
\title{Extract the data file used in a control stream}
\usage{
NMextractDataFile(file, dir.data = NULL)
}
\arguments{
\item{file}{The input control stream or the list file.}

\item{dir.data}{See NMscanInput. If used, only the file name
mentioned in $DATA is used. This file name is looked for in
dir.data. If dir.data is NULL (default), the path is either
absolute (only unix paths starting in / are supported.}
}
\description{
Extract the data file used in a control stream
}
