% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMgenText.R
\name{NMgenText}
\alias{NMgenText}
\title{Generate text for INPUT and possibly DATA sections of NONMEM
control streams.}
\usage{
NMgenText(
  data,
  drop,
  dir.data,
  col.flagn = "FLAG",
  rename,
  pseudo,
  file,
  capitalize = FALSE,
  allow.char.TIME = TRUE,
  quiet
)
}
\arguments{
\item{allow.char.TIME}{For the $DATA text proposal only. Assume
Nonmem can read TIME even if it can't be translated to
numeric. This is necessary if using the 00:00 format. Default
is TRUE.}

\item{nmdir.data}{Deprecated, use
args.NMgenText=list(dir.data="your/path") instead. For the
$DATA text proposal only. The path to the input datafile to be
used in the Nonmem $DATA section. Often, a relative path to
the actual Nonmem run is wanted here.}

\item{nm.copy}{For the $DATA text proposal only. If you plan to
rename columns in Nonmem $DATA, NMwriteData can adjust the
suggested $DATA text. If you plan to use CONC as DV in Nonmem,
you can include nm.rename=c(DV="CONC").}

\item{nm.rename}{For the $DATA text proposal only. If you plan to
rename columns in Nonmem $DATA, NMwriteData can adjust the
suggested $DATA text. If you plan to use BBW instead of BWBASE
in Nonmem, consider nm.rename=c(BWBASE="BBW"). The result is
different from nm.copy since the nm.copy syntax is only
allowed by Nonmem for certain standard column names such as
DV.}

\item{nm.drop}{Only used for generation of proposed text for
Nonmem control stream. Columns to drop in Nonmem $DATA. This
has two implications. One is that the proposed $DATA indicates
=DROP after the given column names. The other that in case it
is a non-numeric column, succeeding columns can still be
included.}

\item{nm.capitalize}{For the $DATA text proposal only. If TRUE,
the suggested text for Nonmem will only contain capital
letters in column names.}
}
\value{
Text for inclusion in Nonmem control stream, invisibly.
}
\description{
The user is provided with text to use in Nonmem. NMwriteSection
can use the results to update the control streams. INPUT lists
names of the data columns while DATA provides a path to data and
ACCEPT/IGNORE statements. Once a column is reached that Nonmem
will not be able to read as a numeric and column is not in
nm.drop, the list is stopped. Only exception is TIME which is not
tested for whether character or not.
}
\seealso{
Other Nonmem: 
\code{\link{NMapplyFilters}()},
\code{\link{NMextractText}()},
\code{\link{NMreadSection}()},
\code{\link{NMwriteData}()},
\code{\link{NMwriteSection}()}
}
\concept{Nonmem}
