% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadCov.R
\name{NMreadCov}
\alias{NMreadCov}
\title{Read in data file}
\usage{
NMreadCov(file, ...)
}
\arguments{
\item{file}{The .cov covariance Nonmem matrix file to read}

\item{...}{Passed to fread}
}
\value{
A matrix with covariance step from NONMEM
}
\description{
Read in data file
}
\details{
This function is taken from nonmem2rx::nmcov which was
    based on function from NMdata.
}
\author{
Philip Delff and Matthew L. Fidler
}
