% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadCtlPars.R
\name{NMreadCtlPars}
\alias{NMreadCtlPars}
\title{Tabulate information from parameter sections in control streams}
\usage{
NMreadCtlPars(lines, section, as.fun)
}
\arguments{
\item{lines}{A control stream as text lines}

\item{section}{The section to read. Typically, "theta", "omega", or "sigma".}

\item{as.fun}{See ?NMscanData}
}
\description{
Tabulate information from parameter sections in control streams
}
\keyword{internal}
