% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadExt.R
\name{NMreadExt}
\alias{NMreadExt}
\title{Read information from Nonmem ext files}
\usage{
NMreadExt(
  file,
  return = "pars",
  as.fun,
  modelname,
  col.model,
  auto.ext,
  tableno = "max",
  file.ext
)
}
\arguments{
\item{file}{Path to the ext file}

\item{return}{The .ext file contains both final parameter
estimates and iterations of the estimates. If
\code{return="pars"} (default) the final estimates are
returned in addition to what other parameter-level information
is found. If \code{return="iterations"}, the iterations are
returned. If \code{return="both"}, both er returned, though in
separate data.frames compiled in a list.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say tibble::as_tibble) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}

\item{modelname}{See ?NMscanData}

\item{col.model}{See ?NMscanData}

\item{auto.ext}{If TRUE (default) the extension will automatically
be modified using `NMdataConf()$file.ext`. This means `file`
can be the path to an input or output control stream, and
`NMreadExt` will still read the `.ext` file.}

\item{tableno}{In case the ext file contains multiple tables, this argument controls which one to choose. The options are
\itemize{

\item "max" (default) Pick the table with the highest table
number. This typically means the results from the last
`$ESTIMATION` step are used.

\item "min" Pick results from the first table available.

\item "all" Keep all results. The tables can be distinguished by
the `tableno` column.

\item an integer greater than 0, in which case the table with this
table number will be picked.
}}

\item{file.ext}{Deprecated. Please use \code{file} instead.}
}
\value{
A list with a final parameter table and a table of the
    iterations
}
\description{
Read information from Nonmem ext files
}
