% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadPhi.R
\name{NMreadPhi}
\alias{NMreadPhi}
\title{Read information from Nonmem phi files}
\usage{
NMreadPhi(file.phi)
}
\arguments{
\item{file.phi}{Path to the phi file}
}
\value{
A list with a final parameter table and a table of the iterations
}
\description{
Read information from Nonmem phi files
}
