% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadShk.R
\name{NMreadShk}
\alias{NMreadShk}
\title{Read Shrinkage data reported by Nonmem}
\usage{
NMreadShk(file, as.fun)
}
\arguments{
\item{file}{A model file. Extension will be replaced by ".shk".}

\item{as.fun}{See ?NMdataConf}
}
\value{
A `data.frame` with shrinkage values, indexes, and name of related parameter, like `OMEGA(1,1)`.
}
\description{
Read Shrinkage data reported by Nonmem
}
\details{
Type 1=etabar
Type 2=Etabar SE
Type 3=P val
Type 4=%Eta shrinkage SD version
Type 5=%EPS shrinkage SD version
Type 6=%Eta shrinkage based on empirical Bayes Variance (SD version)
Type 7=number of subjects used.
Type 8=%Eta shrinkage variance version
Type 9=%Eta shrinkage based on empirical Bayes Variance (variance version)
Type 10=%EPS shrinkage variance version
Type 11=%Relative information
}
