% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadTab.R
\name{NMreadTab}
\alias{NMreadTab}
\title{Read a table file from NONMEM}
\usage{
NMreadTab(file, silent = F, ..., debug = F)
}
\arguments{
\item{file}{path to NONMEM table file}

\item{silent}{logical stating whether or not information is printed.}

\item{...}{Arguments passed to fread.}

\item{debug}{Start by calling browser()?}
}
\value{
Nonmem table as df.
}
\description{
Read a table file from NONMEM
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadCsv}},
  \code{\link{NMscanTables}}, \code{\link{NMtransFilters}},
  \code{\link{NMtransInput}}, \code{\link{NMundoReplace}},
  \code{\link{NMwriteData}}
}
\concept{Nonmem}
