% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreplaceDataFile.R
\name{NMreplaceDataFile}
\alias{NMreplaceDataFile}
\title{Replace data file used in Nonmem control stream}
\usage{
NMreplaceDataFile(files, file.pattern, dir, path.data, newfile = file.mod, ...)
}
\arguments{
\item{files}{Paths to input control streams to modify. See
file.pattern and dir too.}

\item{file.pattern}{A pattern to look for if `dir` is supplied too
(and not `file.mod`). This is used to modify multiple input
control streams at once.}

\item{dir}{Directory in which to look for `file.pattern`. Notice,
use either just `file.mod` or both `dir` and `file.pattern`.}

\item{path.data}{Path to input control stream to use in newfile}

\item{newfile}{A path to a new control stream to write to (and
don't edit contents of `file.mod`). Default is to overwrite `file.mod`.}

\item{...}{Additional arguments to pass to NMwriteSection.}
}
\value{
Lines for a new control stream (invisibly)
}
\description{
Replace data file used in Nonmem control stream
}
\seealso{
Other Nonmem: 
\code{\link{NMapplyFilters}()},
\code{\link{NMextractText}()},
\code{\link{NMgenText}()},
\code{\link{NMreadSection}()},
\code{\link{NMwriteSection}()}
}
\concept{Nonmem}
