% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreplacePart.R
\name{NMreplacePart}
\alias{NMreplacePart}
\title{replace ($)sections of a nonmem control stream}
\usage{
NMreplacePart(path, section, newlines, newpath, backup = T,
  blank.append = T, test = F, debug = F)
}
\arguments{
\item{path}{The run to edit. If a directory is given, the file is assumed to
be called input.txt in that folder.}

\item{section}{The name of the section to update. Example: section="EST" to
edit the sections starting by $EST. See NMgetSection}

\item{newlines}{The new text (including "$SECTION"). Better be broken into
lines in a character vector since this is simply written with
writeLines.}

\item{newpath}{path to new run. If missing, path is used. If NULL, output is
returned rather than written.}

\item{backup}{In case you are overwriting the old file, do you want to
backup the file (to say, backup_input.txt)?}

\item{blank.append}{Append a blank line to output?}

\item{test}{Want to see the resulting input.txt and not write it to disk?
Default is FALSE.}

\item{debug}{start by running browser()?}
}
\description{
Just give the section name, the new lines and the file path, and the
"$section", and the input to Nonmem will be updated.
}
\details{
The new file will be written with unix-style line endings.
}
\examples{
newlines <- "$EST POSTHOC INTERACTION METHOD=1 NOABORT PRINT=5 MAXEVAL=9999 SIG=3"
NMreplacePart(path=NMdata_filepath("examples/nonmem/run001.mod"),
section="EST", newlines=newlines)
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadCsv}},
  \code{\link{NMreadTab}}, \code{\link{NMscanTables}},
  \code{\link{NMtransInput}}, \code{\link{NMundoReplace}},
  \code{\link{NMwriteData}}
}
\concept{Nonmem}
