% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMscanTables.R
\name{NMscanTables}
\alias{NMscanTables}
\title{read all output data tables in nonmem run}
\usage{
NMscanTables(file, details = F, as.fun = NULL, quiet = FALSE, tab.count = TRUE)
}
\arguments{
\item{file}{the nonmem file to read (normally .mod or .lst)}

\item{details}{If TRUE, metadata is added to output. In this case,
you get a list. I'd say, enable if doing programming.}

\item{as.fun}{The default is to return data in data.tables. Pass a
function in as.fun to convert to something else. If
data.frames are wanted, use as.fun=as.data.frame.}

\item{quiet}{The default is to give some information along the way
on what data is found. But consider setting this to TRUE for
non-interactive use.}

\item{tab.count}{Nonmem includes a counter of tables in the
written data files. These are often not useful. However, if
tab.count is TRUE (default), this will be carried forward and
added as a column called TABLENO.}
}
\value{
A list of all the tables as data.frames. If details=TRUE,
    this is in one element, called data, and meta is another
    element. If not, only the element corresponding to data is
    returned.
}
\description{
read all output data tables in nonmem run
}
\seealso{
Other Nonmem: 
\code{\link{NMextractText}()},
\code{\link{NMgetSection}()},
\code{\link{NMreadCsv}()},
\code{\link{NMreadTab}()},
\code{\link{NMtransFilters}()},
\code{\link{NMtransInput}()},
\code{\link{NMwriteData}()}
}
\concept{Nonmem}
