% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMtransFilters.R
\name{NMtransFilters}
\alias{NMtransFilters}
\title{Translate filters in Nonmem and apply to data}
\usage{
NMtransFilters(data, file, text, lines, debug = FALSE)
}
\arguments{
\item{data}{An input data object. Could be read with NMreadCsv or
NMtransInput.}

\item{file}{Path to mod/lst file. Only one of file, text, or lines to be
given. See ?NMgetSection for understanding when to use, file, text, or
lines.}

\item{text}{The mod/lst as characters.}

\item{lines}{The mod/lst as character, line by line.}

\item{debug}{start by running browser()?}
}
\description{
Translate filters in Nonmem and apply to data
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadCsv}},
  \code{\link{NMreadTab}}, \code{\link{NMscanTables}},
  \code{\link{NMtransInput}}, \code{\link{NMundoReplace}},
  \code{\link{NMwriteData}}
}
\concept{Nonmem}
