% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMtransInp.R
\name{NMtransInp}
\alias{NMtransInp}
\title{translate the column names according to the $INPUT section of a control stream}
\usage{
NMtransInp(data, file)
}
\arguments{
\item{data}{the data to translate}

\item{file}{the list file or control stream}
}
\description{
translate the column names according to the $INPUT section of a control stream
}
