% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMtransInput.R
\name{NMtransInput}
\alias{NMtransInput}
\title{read input data and translate names according to the $INPUT section}
\usage{
NMtransInput(file, useRDS = TRUE, dir.data, quiet = FALSE, debug = F)
}
\arguments{
\item{file}{a .lst or a .mod. No matter which you provide, the .mod is
required and is the only one to be read.}

\item{useRDS}{If an rds file is found with the exact same name (except for
.rds instead of say .csv) as the text file mentioned in the Nonmem
control stream, should this be used instead? The default is yes, and
NMwriteData will create this by default too.}

\item{dir.data}{The data directory can only be read from the control stream
(.mod) and not from the output file (.lst). So if you only have the
output file, use dir.data to tell in which directory to find the data file.}

\item{quiet}{Default is to inform a little, but TRUE is useful for
non-interactive stuff.}

\item{debug}{start by running browser()?}
}
\description{
Based on a nonmem run, this function finds the input data and
    reads it. But it reads it like the nonmem run by applying DROP arguments
    and alternative naming of columns in the nonmem run.
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadCsv}},
  \code{\link{NMreadTab}}, \code{\link{NMreplacePart}},
  \code{\link{NMscanTables}}, \code{\link{NMundoReplace}},
  \code{\link{NMwriteData}}
}
\concept{Nonmem}
