% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMundoReplace.R
\name{NMundoReplace}
\alias{NMundoReplace}
\title{Undo NMreplacePart
simply overwrite input.txt with contents of backup_input.txt}
\usage{
NMundoReplace(file, debug = F)
}
\arguments{
\item{file}{The file to recover}

\item{debug}{start by running browser()?}
}
\description{
Undo NMreplacePart
simply overwrite input.txt with contents of backup_input.txt
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadCsv}},
  \code{\link{NMreadTab}}, \code{\link{NMreplacePart}},
  \code{\link{NMscanTables}}, \code{\link{NMtransInput}},
  \code{\link{NMwriteData}}
}
\concept{Nonmem}
