% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSAME.R
\name{NSAME}
\alias{NSAME}
\title{How many SAME reps?}
\usage{
NSAME(x)
}
\arguments{
\item{x}{A character string to test}
}
\description{
How many SAME reps?
}
\examples{
NSAME("SAME")
NSAME(" SAME ")
NSAME("SAME(1)")
NSAME("SAME (1)")
NSAME("  SAME   (  123 ) ")
NSAME("SAME()")
NSAME("SAME (a)")
NSAME("NOT SAME")
}
\keyword{internal}
