% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objInfo.R
\name{objInfo}
\alias{objInfo}
\title{Get stamped info about object}
\usage{
objInfo(data)
}
\arguments{
\item{data}{The dataset to receive info about}
}
\description{
Very simple function that extracts info stored by stampObj.
}
\seealso{
stampObj

Other DataCreate: 
\code{\link{NMorderColumns}()},
\code{\link{findCovs}()},
\code{\link{findVars}()},
\code{\link{flagsAssign}()},
\code{\link{flagsCount}()},
\code{\link{mergeCheck}()},
\code{\link{stampObj}()},
\code{\link{tmpcol}()}
}
\concept{DataCreate}
