% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot raw nonmem data}
\usage{
plot_data(database, time = "TAD", dv = "DV")
}
\arguments{
\item{database}{name of the database}

\item{time}{Name of the time column you want on the x-axis (e.g. 'TAD')}

\item{dv}{Dependend variable (e.g. concentration 'DV')}
}
\description{
Plot raw nonmem data
}
