% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_NMdata.R
\name{print.summary_NMdata}
\alias{print.summary_NMdata}
\title{print method for NMdata summaries}
\usage{
\method{print}{summary_NMdata}(x, ...)
}
\arguments{
\item{x}{The summary object to be printed. See ?summary.NMdata}

\item{...}{Arguments passed to other print methods.}
}
\description{
print method for NMdata summaries
}
