% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.NMdata.R
\name{rbind.NMdata}
\alias{rbind.NMdata}
\title{combine NMdata objects
import data.table}
\usage{
\method{rbind}{NMdata}(...)
}
\arguments{
\item{...}{The NMdata objects to combine}
}
\value{
If the first table in the first argument is a data.table,
    unkeyed data.tables will be returned. If not, data.frames will
    be returned.
}
\description{
combine NMdata objects
import data.table
}
\details{
Each element in a NMdata object is a data.frame or
    data.table. The name of the element represents the variability
    level represented in the data.frame. Hence, this function
    rbinds elements at equal varibility level, say obj1$row and
    obj2$row.
}
