% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTables.R
\name{reduceTables}
\alias{reduceTables}
\title{reduce tables from NMscanTables to fewer objects}
\usage{
reduceTables(tables, col.nmout)
}
\arguments{
\item{tables}{Object from NMtables}

\item{col.nmout}{The name of the column holding logical of whether
row was found in output tables.}
}
\value{
A list of data.tables: One row-level, one id-level data
    set and their column specifications.
}
\description{
reduce tables from NMscanTables to fewer objects
}
\details{
This function is under development. It is not polished
    for other than internal use by NMscanData. It will likely
    return a different format in the future.
}
\keyword{internal}
