% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameByContents.R
\name{renameByContents}
\alias{renameByContents}
\title{Rename columns matching properties of data contents}
\usage{
renameByContents(data, fun.test, fun.rename, invert.test = FALSE, as.fun)
}
\arguments{
\item{data}{data.frame in which to rename columns}

\item{fun.test}{Function that returns TRUE for columns to be
renamed.}

\item{fun.rename}{Function that takes the existing column name and
returns the new one.}

\item{invert.test}{Rename those where FALSE is returned from
fun.test.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say tibble::as_tibble) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}
}
\value{
data with (some) new column names. Class as defined by
    as.fun.
}
\description{
For instance, lowercase all columns that Nonmem
    cannot interpret (as numeric).
}
