% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_by_id.R
\name{seq_by_id}
\alias{seq_by_id}
\title{Create sequences for NONMEM dataset}
\usage{
seq_by_id(id_col)
}
\arguments{
\item{id_col}{Values from the (patient) ID column}

\item{occ_col}{Values from the occasion (OCC) column}
}
\description{
Create sequences for NONMEM dataset
}
\keyword{seq_by_id}
