% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadParsText.R
\name{splitFields}
\alias{splitFields}
\title{splitFields splits the fields format string into the splitters
and the variable names. NB, this is interpreting the user-provided
fields, it is not looking at control stream text,}
\usage{
splitFields(format, spaces.split = FALSE)
}
\description{
splitFields splits the fields format string into the splitters
and the variable names. NB, this is interpreting the user-provided
fields, it is not looking at control stream text,
}
\keyword{internal}
