% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdata_utils.R
\name{summary.NMdata}
\alias{summary.NMdata}
\title{print method for NMdata objects}
\usage{
\method{summary}{NMdata}(data, ...)
}
\description{
print method for NMdata objects
}
\details{
The subjects are counted conditioned on the nmout
    column. If only id-level output tables are present, there are
    no nmout=TRUE rows. This means that in this case it will
    report that no IDs are found in output. The correct statement
    is that records are found for zero subjects in output tables.
}
