% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Create table one}
\usage{
table1(db, id = "ID")
}
\arguments{
\item{db}{Database containing patient characteristics}

\item{id}{Name of ID column}
}
\description{
Create table one
}
