% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{tad}
\alias{tad}
\title{Calculate time after dose}
\usage{
tad(
  database,
  date = "DATE",
  time = "TIME",
  dose = "DOSE",
  id = "ID",
  format = "\%m/\%d/\%Y \%H:\%M"
)
}
\arguments{
\item{date}{The name of the date column (default is 'DATE') Should be as characters}

\item{time}{The name of the time column (default is 'TIME') Should be as characters}

\item{dose}{The name of the dose column (default is 'DOSE')}

\item{format}{The format of the date and time column (default "\%m/\%d/\%Y \%H:\%M" )}
}
\description{
A function to take the columns with calendar date and clock times and uses
this to calculate the time after dose. Calendar date and clock times should
be as characters. The format has a default but can be provided as argument.
If the format is not correct it cannot calculate the time after dose and will
generate a lot of NA's.
}
\keyword{TAD}
