% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpcol.R
\name{tmpcol}
\alias{tmpcol}
\title{generate a name for a new data column that is not already in use.}
\usage{
tmpcol(
  data,
  names = NULL,
  base = "atmpcol999",
  max.it = 100,
  prefer.plain = TRUE
)
}
\arguments{
\item{data}{The dataset to find a new element name for}

\item{names}{Character vector of names that must not be matched. Only one of
data and names can be supplied.}

\item{base}{The base name of the new element. A number will appended to this
string that will ensure that the new element name is not already in use.}

\item{max.it}{Maximum number of iterations on element name.}

\item{prefer.plain}{I base isn't in use already, use it without a digit
appended?}
}
\description{
generate a name for a new data column that is not already in use.
}
\seealso{
Other DataWrangling: 
\code{\link{NMorderColumns}()},
\code{\link{NMscanData}()},
\code{\link{findCovs}()},
\code{\link{findVars}()},
\code{\link{mergeCheck}()}
}
\concept{DataWrangling}
