% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadCtlPars.R
\name{triagSize}
\alias{triagSize}
\title{Calculate number of elements for matrix specification}
\usage{
triagSize(diagSize)
}
\description{
calculate number of elements in the diagonal and lower triangle of
a matrix, based on the length of the diagonal
}
\keyword{internal}
