% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdata_utils.R
\name{unNMdata}
\alias{unNMdata}
\title{Remove NMdata class and discard NMdata meta data}
\usage{
unNMdata(x)
}
\arguments{
\item{x}{An 'NMdata' object.}
}
\value{
x stripped from the 'NMdata' class
}
\description{
Remove NMdata class and discard NMdata meta data
}
