% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNMinfo.R
\name{writeNMinfo}
\alias{writeNMinfo}
\title{Do the actual writing of meta data}
\usage{
writeNMinfo(data, meta, append = FALSE, byRef = TRUE)
}
\arguments{
\item{data}{A data set}

\item{meta}{The meta data to attach}

\item{append}{If FALSE, the existing meta data will be removed. If
TRUE, metadata will be appended to existing metadata. However,
this will not work recursively.}

\item{byRef}{Should always be TRUE.}
}
\value{
The data with meta data attached
}
\description{
Do the actual writing of meta data
}
\keyword{internal}
