% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTextFile.R
\name{writeTextFile}
\alias{writeTextFile}
\title{Conveniently write text lines to file}
\usage{
writeTextFile(lines, file)
}
\arguments{
\item{lines}{the character lines to write}

\item{file}{The file name path to write to}
}
\description{
Conveniently write text lines to file
}
\keyword{internal}
