% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCor.R
\name{addOmegaCorr}
\alias{addOmegaCorr}
\title{Deprecated: use addCor. Add correlations to parameter table}
\usage{
addOmegaCorr(...)
}
\arguments{
\item{...}{Passed to addCor}
}
\description{
Anything arguments are passed to `addCor()`. See `?addCor()`.
}
