% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadInits.R
\name{addSameBlocks}
\alias{addSameBlocks}
\title{Create a variable in inital value table to keep track of SAME
blocks i.e. parameters that are part of a single distribution}
\usage{
addSameBlocks(inits)
}
\arguments{
\item{inits}{Table of initial values as created by NMreadInits().}
}
\description{
Create a variable in inital value table to keep track of SAME
blocks i.e. parameters that are part of a single distribution
}
\details{
sameblock:

if not part of a distribution repeated using SAME: 0

if part of a distribution repeated using SAME: counter (1,2,...)
of the unique distribution blocks that are being reused.

Nsameblock: The number of SAME calls used for a distribution
block. If SAME(N) notation is used, Nsameblock=N.
}
\author{
Brian Reilly
}
\keyword{internal}
