% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTAD.R
\name{addTAD}
\alias{addTAD}
\title{Add cumulative number of doses, time of last dose and time since last dose to data}
\usage{
addTAD(
  data,
  col.time = "TIME",
  col.tdos = "TDOS",
  col.tad = "TAD",
  col.ndoses = "NDOSES",
  col.evid = "EVID",
  col.amt = "AMT",
  subset.dos,
  subset.is.complete,
  by = "ID",
  as.fun
)
}
\arguments{
\item{data}{The data set to add the variables to.}

\item{col.time}{Name of time column (created by addTAD). Default
it TIME.}

\item{col.tdos}{Name of the time of previous dose column (created
by addTAD). Default is TDOS.}

\item{col.tad}{Name of the time of prvious dose column (created by
addTAD). Default is TAD.}

\item{col.ndoses}{The name of the column (created by addTAD) that
holds the cumulative number of doses administered to the
subject.}

\item{col.evid}{The name of the event ID column. This must exist
in data. Default is EVID.}

\item{col.amt}{col.evid The name of the dose amount column. This
must exist in data. Default is AMT.}

\item{subset.dos}{A string that will be evaluated as a custom
expression to identify relevant events. See subset.is.complete
as well.}

\item{subset.is.complete}{Only used in combination with
non-missing subset.dos. By default, subset.dos is used in
addition to the impact of col.evid and col.amt. If
subset.is.complete=TRUE, subset.dos is used alone, and
col.evid and col.amt are completely ignored. This is typically
useful if the events are not doses but other events that are
not expressed as a typical dose combination of EVID and AMT
columns.}

\item{by}{Columns to do calculations within. Default is ID.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say tibble::as_tibble) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}
}
\description{
For now, doses have to be in data as EVID=1 and/or EVID=4
records. They can be in the format of one row per dose or repeated
dosing notation using ADDL and II.
}
\details{
This is still experimental.
}
