% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colLabels.R
\name{colLabels}
\alias{colLabels}
\title{Extract column labels as defined in SAS}
\usage{
colLabels(x, sort = "alpha")
}
\arguments{
\item{x}{object with elements containing label attributes.}

\item{sort}{If sort="alpha", results are sorted alphabetically.}
}
\value{
A data.frame with variable and their labels
}
\description{
Extract column labels as defined in SAS
}
\seealso{
compareCols NMinfo
}
