% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colLabels.R
\name{colLabels}
\alias{colLabels}
\title{Extract column labels as defined in SAS}
\usage{
colLabels(...)
}
\arguments{
\item{...}{See `?compareCols`}
}
\description{
Extract column labels as defined in SAS
}
\seealso{
compareCols NMinfo
}
