% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcastSe.R
\name{dcastSe}
\alias{dcastSe}
\title{A standard-evaluation interface to `data.table::dcast()}
\usage{
dcastSe(data, l, r, as.fun, ...)
}
\arguments{
\item{data}{data set to transpose (widen)}

\item{l}{left-hand side variables as character vector. Result will
be long/vertical in these variables.}

\item{r}{left-hand side variables as character vector. Result will
be wide in these variables.}

\item{as.fun}{An optional function to convert results with. If `data` is a `data.table`, the default is to return a `data.table`, and if not the default is to return a `data.frame`.}

\item{...}{Additional arguments passed to `data.table::dcast()`.}
}
\value{
a data.table
}
\description{
data.table's dcast() transforms from long to wide format. It lacks a standard-evalutation interface for programming, and dcastSe() is an attempt to fill that gap, still using data.table::dcast() to perform the operation.
}
\details{
Only two of l, r, and value.var are needed (see ?data.table::dcast for value.var), and the others will be derived. However, if value.var is a function, both l and r are needed.
}
