% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims}
\alias{dims}
\title{Get dimensions of multiple objects}
\usage{
dims(..., list.data, keep.names = TRUE, as.fun = NULL, keepNames)
}
\arguments{
\item{...}{data sets}

\item{list.data}{As alternative to ..., you can supply the data
sets in a list here.}

\item{keep.names}{If TRUE, the original dataset names are used in
reported table. If not, generic x1, x2,... are used. The
latter may be preferred for readability in some cases.}

\item{as.fun}{A function that will be run on the result before
returning. If first input data set is a data.table, the
default is to return a data.table, if not the default is to
return a data.frame. Use whatever to get what fits in with
your workflow. Default can be configured with NMdataConf.}

\item{keepNames}{Deprecated. Use keep.names instead.}
}
\value{
A data.frame with dimensions of objects in ... Actual
    class defined by as.fun.
}
\description{
Get dimensions of multiple objects
}
\seealso{
Other DataWrangling: 
\code{\link{compareCols}()},
\code{\link{listMissings}()}
}
\concept{DataWrangling}
