% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filePathSimple.R
\name{filePathSimple}
\alias{filePathSimple}
\title{Tidy paths}
\usage{
filePathSimple(..., debug = F)
}
\arguments{
\item{...}{additional arguments passed to file.path().}

\item{debug}{start by running browser()?}
}
\description{
Used to tidy up paths. Passes arguments to
    file.path(). The function is intended to return a canonical
    path format, i.e. paths that can be compared by simple string
    comparisson.
}
\concept{FileSystem}
