% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCovs.R
\name{findCovs}
\alias{findCovs}
\title{Extract columns that do not vary within variables in a data.frame}
\usage{
findCovs(data, cols.id = NULL, debug = F)
}
\arguments{
\item{data}{data.frame in which to look for covariates}

\item{cols.id}{covariates will be searched for in combinations of values in
these columns. Often cols.id will be either empty or ID. But it
can also be both say c("ID","DRUG") or c("ID","TRT").}

\item{debug}{start by running browser()?}
}
\description{
Extract columns that do not vary within variables in a data.frame
}
\examples{
\dontrun{
dat <- NMreadTab("TABLE_OUTPUT.txt")
### very common use
findCovs(dat,cols.id="ID")
###  an ID column is not needed.
findCovs(dat,cols.id=c())
### need a new data.frame to test for length(cols.id)>1
}
}
\seealso{
Other DataWrangling: 
\code{\link{NMorderColumns}()},
\code{\link{NMscanData}()},
\code{\link{findVars}()},
\code{\link{mergeCheck}()},
\code{\link{tmpcol}()}
}
\concept{DataWrangling}
