% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVars.R
\name{findVars}
\alias{findVars}
\title{Extract columns that vary within values of other columns in a data.frame}
\usage{
findVars(data, cols.id = NULL, debug = F)
}
\arguments{
\item{data}{data.frame in which to look for covariates}

\item{cols.id}{optional covariates will be searched for in
combinations of values in these columns. Often cols.id will be
either empty or ID. But it can also be both say c("ID","DRUG")
or c("ID","TRT").}

\item{debug}{start by running browser()?}
}
\description{
Extract columns that vary within values of other columns in a data.frame
}
\details{
Use this to exclude columns that are constant within
    cols.id. If cols.id=ID, this could be to get only time-varying
    covariates.
}
\seealso{
Other DataWrangling: 
\code{\link{NMorderColumns}()},
\code{\link{NMscanData}()},
\code{\link{findCovs}()},
\code{\link{mergeCheck}()},
\code{\link{tmpcol}()}
}
\concept{DataWrangling}
