% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagsAssign.R
\name{flagsAssign}
\alias{flagsAssign}
\title{Assign exclusion flags to a dataset based on specified table}
\usage{
flagsAssign(data, tab.flags, return.all = F, LLOQ = NULL,
  col.id = "ID", col.dv = "DV", debug = F)
}
\arguments{
\item{data}{The dataset to assign flags to.}

\item{tab.flags}{A data.frame containing at least these named columns: FLAG,
flag, condition. Condition is disregarded for FLAG==0.}

\item{return.all}{If TRUE, both the edited dataset and the table of flags
are returned. If FALSE (default) only the edited dataset is returned.}

\item{LLOQ}{At the moment a list with FLAG, condition, and LLOQ (value). If
it does not contain "flag", flag will be set to "Value below LLOQ". LLOQ
can only handle one value of LLOQ. This is insuffiecient for PKPD
datasets. Likely to change.}

\item{col.id}{The name of the subject ID column. Default is "ID".}

\item{col.dv}{The name of the data value column. Default is "DV".}

\item{debug}{Start by calling browser()?}
}
\value{
The dataset with flags added. See parameter flags.return as well.
}
\description{
The aim with this function is to take a (say PK) dataset and a pre-specified
table of flags, assign the flags automatically.
}
