% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagsCount.R
\name{flagsCount}
\alias{flagsCount}
\title{Create an overview of number of retained and discarded datapoints.}
\usage{
flagsCount(data, tab.flags, file, col.id = "ID", by = NULL,
  debug = F)
}
\arguments{
\item{data}{The dataset including both FLAG and flag columns.}

\item{tab.flags}{A data.frame containing at least these named
columns: FLAG, flag, condition. Condition is disregarded for
FLAG==0.}

\item{file}{A file to write the table of flag counts to. Will
probably be removed and put in a separate function.}

\item{col.id}{The name of the subject ID column. Default is
"ID".@param col.id The name of the subject ID column. Default
is "ID".}

\item{by}{An optional column to group the counting by. This could
be "STUDY", "DRUG", "EVID", or a combination of multiple
columns.}

\item{debug}{Start by calling browser()?}
}
\value{
A summary table with number of discarded and retained
    subjects and observations when applying each condition in the
    flag table. "discarded" means that the reduction of number of
    observations and subjects resulting from the flag, "retained"
    means the numbers that are left after application of the
    flag. The default is "both" which will report both.
}
\description{
Create an overview of number of retained and discarded datapoints.
}
\details{
Notice number of subjects in N.discarded mode can be
    misunderstood. If two is reported, it can mean that the
    remining one observation of these two subjects are discarded
    due to this flag. The majority of the samples can have been
    discarded by earlier flags.
}
