% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnExtension.R
\name{fnExtension}
\alias{fnExtension}
\title{Change file name extension}
\usage{
fnExtension(fn, ext)
}
\arguments{
\item{fn}{file name}

\item{ext}{new file name extension}
}
\value{
A text string
}
\description{
Very simple but often applicable function to change the file name extension (from say file.lst to file.mod)
}
\examples{
fnExtension("file.lst",".mod")
fnExtension("file.lst","")
}
