% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnExtension.R
\name{fnExtension}
\alias{fnExtension}
\title{Change file name extension}
\usage{
fnExtension(fn, ext)
}
\arguments{
\item{fn}{file name. Often ending in an extension after a period
but the extension is not needed.}

\item{ext}{new file name extension. If omitted or NULL, the
extension of fn is returned.}
}
\value{
A text string
}
\description{
Very simple but often applicable function to retrieve or change
the file name extension (from say file.lst to file.mod)
}
\examples{
fnExtension("file.lst",".mod")
fnExtension("file.lst","mod")
fnExtension("file.lst","..mod")
fnExtension("file.lst",cc(.mod,xml))
fnExtension(cc(file1.lst,file2.lst),cc(.xml))
fnExtension(cc(file1.lst,file2.lst),cc(.xml,.cov))
fnExtension("file.lst","")
fnExtension("file.lst")
}
