% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgs.R
\name{getArgs}
\alias{getArgs}
\title{Get provided arguments as a named list}
\usage{
getArgs(call, env)
}
\arguments{
\item{call}{Function call as provided by \code{sys.call()}.}

\item{env}{Environment in which to evaluate the arguments.}
}
\value{
A named list of arguments and their values
}
\description{
Get provided arguments as a named list
}
\examples{
afun <- function(){
NMdata:::getArgs(sys.call(),parent.frame())
}
afun()
}
\seealso{
Other arguments: 
\code{\link{deprecatedArg}()}
}
\concept{arguments}
\keyword{internal}
