% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSAME.R
\name{isSAME}
\alias{isSAME}
\title{Test whether an initial value is of type SAME}
\usage{
isSAME(x)
}
\arguments{
\item{x}{A character string to test}
}
\description{
Test whether an initial value is of type SAME
}
\examples{
isSAME("SAME")
isSAME(" SAME ")
isSAME("SAME(1)")
isSAME("SAME (1)")
isSAME("  SAME   (  123 ) ")
isSAME("SAME()")
isSAME("SAME (a)")
isSAME("NOT SAME")
}
\keyword{internal}
