% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{load.data}
\alias{load.data}
\title{function to import data file based on file type (either 'xlsx' or 'csv')}
\usage{
load.data(name, file_type)
}
\arguments{
\item{name}{file name without the extension}

\item{file_type}{Type of file xlsx or csv}
\item{na_value}{optional argument to give the value that specifies NA}
}
\description{
function to import data file based on file type (either 'xlsx' or 'csv')
}
