% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{loaddata}
\alias{loaddata}
\title{function to import data file based on file type (either 'xlsx' or 'csv')}
\usage{
loaddata(name, file_type, na_value)
}
\arguments{
\item{name}{file name without extension}

\item{file_type}{type of file xlsx or csv}

\item{na_value}{optional argument to give the value that specifies NA}
}
\description{
function to import data file based on file type (either 'xlsx' or 'csv')
}
\examples{
-
}
\keyword{csv}
\keyword{xlsx}
