% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdata_utils.R
\name{messageWrap}
\alias{messageWrap}
\title{Pretty wrapping of lines in NMdata vignettes}
\usage{
messageWrap(..., fun.msg = message, prefix = "\\n", initial = "", width = 85)
}
\arguments{
\item{...}{parameters to pass to strwrap}

\item{prefix}{Passed to strwrap. Default is " ".}

\item{initial}{Passed to strwrap. Default is an empty string.}

\item{width}{Passed to strwrap. Default is 80.}
}
\description{
Pretty wrapping of lines in NMdata vignettes
}
