% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageWrap.R
\name{messageWrap}
\alias{messageWrap}
\title{Pretty wrapping of lines in NMdata vignettes}
\usage{
messageWrap(
  ...,
  fun.msg = message,
  prefix = "\\n",
  initial = "",
  width,
  track.msg = FALSE
)
}
\arguments{
\item{...}{parameters to pass to strwrap}

\item{fun.msg}{The function to pass the text through. Typically,
message, warning, or stop. If NULL, nothing will happen, and
NULL is invisibly returned.}

\item{prefix}{Passed to strwrap. Default is "\\n".}

\item{initial}{Passed to strwrap. Default is an empty string.}

\item{width}{Passed to strwrap. Default is 80.}

\item{track.msg}{If TRUE, the name of the function throwing the
message/warning/error is mentioned. This is not default but
useful when using function inside other functions.}
}
\value{
Nothing.
}
\description{
Pretty wrapping of lines in NMdata vignettes
}
\keyword{internal}
