% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameMultipleFiles.R, R/nameMultipleModels.R
\name{nameMultipleFiles}
\alias{nameMultipleFiles}
\title{Create file names for multiple list elements}
\usage{
nameMultipleFiles(fn, list.obj)

nameMultipleFiles(fn, list.obj)
}
\arguments{
\item{fn}{A file name to use as the stem for new file names}

\item{list.obj}{List of objects to create file names for}
}
\description{
Can currently only append a counter.
}
\keyword{internal}
