% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcheckData.R
\name{NMcheckData}
\alias{NMcheckData}
\title{Check data for Nonmem compatibility}
\usage{
NMcheckData(
  data,
  col.id = "ID",
  col.time = "TIME",
  col.flagn,
  col.row = NULL,
  na.strings,
  as.fun
)
}
\arguments{
\item{data}{The data to check. data.frame, data.table, tibble,
anything that can be converted to data.table.}

\item{col.id}{The name of the column that holds the subject
identifier. Default is "ID".}

\item{col.time}{The name of the column holding actual time.}

\item{col.flagn}{Optionally, the name of the column holding
numeric exclusion flags. Default value is FLAG and can be
configured using NMdataConf. Disable by using col.flagn=FALSE.}

\item{col.row}{A column with a unique value for each row. Such a
column is recommended to use if possible. Default ("ROW") can
be modified using NMdataConf.}

\item{na.strings}{Strings to be accepted when trying to convert
characters to numerics. This will typically be a string that
represents missing values. Default is "." even though most
users will use actual NA (NA_character), i.e. not a
string. See ?NMisNumeric.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say tibble::as_tibble) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}
}
\description{
Check data in various ways for compatibility with Nonmem. Some
findings will be reported even if they will not make Nonmem fail
but because they are typical dataset issues.
}
\details{
The following checks are performed. The term "numeric"
    does not refer to a numeric representation in R, but
    compatibility with Nonmem. The character string "2" is in this
    sense a valid numeric, "id2" is not.  \itemize{ \item Column
    names must be unique and not contain special characters

\item If an exclusion flag is used (for ACCEPT/IGNORE in Nonmem),
    elements must be non-missing and integers. If an exclusion
    flag is found, the rest of the checks are performed on rows
    where that flag equals 0 (zero) only.

\item col.time (TIME),
    EVID, ID, CMT, MDV: If present, elements must be non-missing
    and numeric.

\item col.time (TIME) must be non-negative

\item EVID must be in {0,1,2,3,4}

\item CMT must be positive integers. However, can be missing or zero for EVID==3.

\item MDV must be the binary (1/0) representation of is.na(DV)

\item AMT must be 0 or NA for EVID 0 and 2

\item AMT must be positive for EVID 1 and 4

\item DV must be numeric

\item DV must be missing for EVID in {1,4}.

\item ID must be positive and values cannot be disjoint (all
    records for each ID must be following each other. This is
    technically not a requirement in Nonmem but most often an
    error. Use a second ID column if you deliberately want to
    soften this check)

\item TIME cannot be decreasing within ID, unless EVID in {3,4}.

\item all ID's must have doses (EVID in {1,4})

\item all ID's must have observations (EVID==0)

\item If a unqique row identifier is used, this must be
    non-missing, increasing, integer

\item Character values must not contain commas (they will mess up
    writing/reading csv) }
}
