% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcheckDataFile.R
\name{NMcheckDataFile}
\alias{NMcheckDataFile}
\title{check input data based on control stream}
\usage{
NMcheckDataFile(
  file,
  col.row,
  col.id = "ID",
  use.rds = FALSE,
  quiet = FALSE,
  file.mod,
  as.fun,
  ...
)
}
\arguments{
\item{file}{a model file (input or output control stream)}

\item{col.row}{row identifier}

\item{col.id}{subject identifier}

\item{use.rds}{use rds file instead of csv. This is typically not wanted because we want to test the input data as read by NONMEM.}

\item{quiet}{Keep quiet? Default is FALSE.}

\item{file.mod}{How to find the input control stream if you are using the output control stream.}

\item{as.fun}{The function to run results through before returning them.}

\item{...}{passed to NMcheckData}
}
\description{
Finds input data and checks compatibility with nonmem control
stream and runs NMcheckData. Don't call this function directly -
use the file argument in NMcheckData instead.
}
\keyword{internal}
