% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcode2R.R
\name{NMcode2R}
\alias{NMcode2R}
\title{Translate Nonmem $PK, $PRED sections or other Nonmem code to R code}
\usage{
NMcode2R(text)
}
\arguments{
\item{text}{the Nonmem code.}
}
\value{
R code as text
}
\description{
Translate Nonmem $PK, $PRED sections or other Nonmem code to R code
}
\details{
You probably want to run this on text obtained by using
    the NMreadSection function.
}
\keyword{internal}
