% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMextractDataFile.R
\name{NMextractDataFile}
\alias{NMextractDataFile}
\title{Extract the data file used in a control stream}
\usage{
NMextractDataFile(file, dir.data = NULL)
}
\arguments{
\item{file}{The input control stream or the list file.}

\item{dir.data}{See NMscanInput. If used, only the file name
mentioned in $DATA is used. dir.data will be used as the path,
and the existence of the file in that directory is not
checked.}
}
\value{
The path to the input data file.
}
\description{
Extract the data file used in a control stream
}
