% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMisNumeric.R
\name{NMisNumeric}
\alias{NMisNumeric}
\title{Test if a variable can be interpreted by Nonmem}
\usage{
NMisNumeric(x)
}
\arguments{
\item{x}{The vector to check Don't export}
}
\value{
TRUE or FALSE
}
\description{
Nonmem can only interpret numeric data. However, a
    factor or a character variable may very well be interpretable
    by Nonmem (e.g. "33"). This function tells whether Nonmem will
    be able to read it.
}
