% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMorderColumns.R
\name{NMorderColumns}
\alias{NMorderColumns}
\title{Order columns in dataset for use in Nonmem.}
\usage{
NMorderColumns(
  data,
  first,
  last,
  lower.last = FALSE,
  chars.last = TRUE,
  nomtime = "NOMTIME",
  row = "ROW",
  as.fun = NULL
)
}
\arguments{
\item{data}{The dataset which columns to reorder.}

\item{first}{Columns that should come almost first. See details.}

\item{last}{Columns to move to back of dataset. If you work with a
large dataset, and some columns are irrelevant for the Nonmem
runs, you can use this argument.}

\item{lower.last}{Should columns which names contain lowercase
characters be moved towards the back? Some people use a
standard of lowercase variables (say "race") being character
representations ("asian", "caucasian", etc.) variables and the
uppercase (1,2,...) being the numeric representation for
Nonmem.}

\item{chars.last}{Should columns which cannot be converted to
numeric be put towards the end? A column can be a character or
a factor in R, but still be valid in Nonmem (often the case
for ID which can only contain numeric digits but really is a
character or factor). So rather than only looking at the
column class, the columns are attempted converted to
numeric. Notice, it will attempted to be converted to numeric
to test whether Nonmem will be able to make sense of it, but
the values in the resulting dataset will be untouched. No
values will be edited. If TRUE, logicals will always be put
last. NA's must be NA or ".".}

\item{nomtime}{The name of the column containing nominal time. If
given, it will put the column quite far left, just after row
counter and ID.}

\item{row}{A row counter column. This will be the first column in
the dataset. Technically, you can use it for whatever column
you want first.}

\item{as.fun}{The default is to return data in a data.table. Pass
a function in as.fun to convert to something else. If
data.frames are wanted, use as.fun=as.data.frame.}
}
\description{
Order data columns for easy export to Nonmem. No data values are
edited. The order is configurable through multiple arguments. See
details.
}
\details{
This function will change the order of columns but it
    will never edit values in any columns.

The ordering is by the following steps, each step depending on
corresponding argument. Not
\itemize{
 \item{"row - "}{Row id if argument row is non-NULL}
 \item{"not editable - "}{ID (if a column is called ID)}
 \item{"nomtime - "}{Nominal time.}
 \item{"first - "}{user-specified first columns}
 \item{"not editable - "}{Standard Nonmem columns: TIME, EVID, CMT, AMT, RATE, DV, MDV}
 \item{"last - "}{user-specified last columns}
 \item{"chars.last - "}{numeric, or interpretable as numeric}
 \item{"not editable - "}{less often used nonmem names: FLAG, OCC, ROUTE, GRP, TRIAL, DRUG, STUDY}
 \item{"lower.last - "}{lower case in name}
 \item{"not editable - "}{Alphabetic/numeric sorting}
}
}
\seealso{
Other DataCreate: 
\code{\link{findCovs}()},
\code{\link{findVars}()},
\code{\link{mergeCheck}()},
\code{\link{tmpcol}()}
}
\concept{DataCreate}
