% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadCov.R
\name{NMreadCov}
\alias{NMreadCov}
\title{Read in data file}
\usage{
NMreadCov(file, auto.ext, ...)
}
\arguments{
\item{file}{The .cov covariance Nonmem matrix file to read}

\item{auto.ext}{If `TRUE` (default) the extension will automatically
be modified using `NMdataConf()$file.cov`. This means `file`
can be the path to an input or output control stream, and
`NMreadCov()` will still read the `.cov` file.}

\item{...}{Passed to fread}
}
\value{
A matrix with covariance step from NONMEM
}
\description{
Read in data file
}
\details{
This function is taken from nonmem2rx::nmcov which was
    based on NMdata::NMreadTab.
}
\author{
Philip Delff and Matthew L. Fidler
}
