% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadCsv.R
\name{NMreadCsv}
\alias{NMreadCsv}
\title{read csv data that is written to be read by nonmem}
\usage{
NMreadCsv(file, na.strings = ".", header = TRUE,
  stringsAsFactors = FALSE, ...)
}
\arguments{
\item{file}{The file to read. Must be pure text.}

\item{na.strings}{See read.table}

\item{header}{See read.table.}

\item{stringsAsFactors}{See read.table.}

\item{...}{passed to read.csv}
}
\description{
read csv data that is written to be read by nonmem
}
\details{
This is just a shortcut to read.csv so you don't have to remember
    how to read the data that was exported for nonmem.
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadTab}},
  \code{\link{NMreplacePart}}, \code{\link{NMscanTables}},
  \code{\link{NMtransInput}}, \code{\link{NMundoReplace}},
  \code{\link{NMwriteData}}
}
\concept{Nonmem}
