% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadCsv.R
\name{NMreadCsv}
\alias{NMreadCsv}
\title{read csv data that is written to be read by nonmem}
\usage{
NMreadCsv(file, na.strings = ".", header = TRUE,
  stringsAsFactors = FALSE, as.dt = TRUE, ...)
}
\arguments{
\item{file}{The file to read. Must be pure text.}

\item{na.strings}{See read.table}

\item{header}{See read.table.}

\item{stringsAsFactors}{See read.table.}

\item{as.dt}{Return a data.table? data.table is default, if not a data.frame
is returned.}

\item{...}{passed to read.csv}
}
\description{
read csv data that is written to be read by nonmem
}
\details{
This is just a shortcut to fread so you don't have to remember how
    to read the data that was exported for nonmem.
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadTab}},
  \code{\link{NMscanTables}}, \code{\link{NMtransFilters}},
  \code{\link{NMtransInput}}, \code{\link{NMwriteData}}
}
\concept{Nonmem}
