% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadCsv.R
\name{NMreadCsv}
\alias{NMreadCsv}
\title{read csv data that is written to be read by nonmem}
\usage{
NMreadCsv(
  file,
  na.strings = ".",
  header = TRUE,
  stringsAsFactors = FALSE,
  as.fun = NULL,
  ...
)
}
\arguments{
\item{file}{The file to read. Must be pure text.}

\item{na.strings}{See read.table}

\item{header}{See read.table.}

\item{stringsAsFactors}{See read.table.}

\item{as.fun}{The default is to return data as a data.frame. Pass a function
(say tibble::as_tibble) in as.fun to convert to something else. If
data.tables are wanted, use as.fun="data.table". The default can be
configured using NMdataConf.}

\item{...}{passed to read.csv}
}
\description{
This function is especially useful if the csv file was written
using NMwriteData.
}
\details{
This is just a shortcut to fread so you don't have to remember how
    to read the data that was exported for nonmem.
}
\seealso{
NMwriteData

Other DataRead: 
\code{\link{NMreadTab}()},
\code{\link{NMscanData}()},
\code{\link{NMscanInput}()},
\code{\link{NMscanTables}()}
}
\concept{DataRead}
