% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadCtlPars.R
\name{NMreadCtlPars}
\alias{NMreadCtlPars}
\title{Tabulate information from parameter sections in control streams}
\usage{
NMreadCtlPars(file, lines, section, as.fun)
}
\arguments{
\item{file}{Path to a control stream. See `lines` too.}

\item{lines}{A control stream as text lines. Use this or `file`.}

\item{section}{The section to read. Typically, "theta", "omega",
or "sigma". Default is those three.}

\item{as.fun}{See ?NMscanData}
}
\description{
Tabulate information from parameter sections in control streams
}
\keyword{internal}
