% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadCtlPars.R
\name{NMreadCtlPars}
\alias{NMreadCtlPars}
\title{Identify active elements in a parameter section}
\usage{
NMreadCtlPars(lines, section, as.fun)
}
\arguments{
\item{lines}{A control stream as text lines}

\item{section}{The section to read. Typically, "theta", "omega", or "sigma".}

\item{as.fun}{See ?NMscanData}
}
\description{
Identify active elements in a parameter section
}
\keyword{internal}
