% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadExt.R
\name{NMreadExt}
\alias{NMreadExt}
\title{Read information from Nonmem ext files}
\usage{
NMreadExt(file.ext, return = "pars", as.fun, modelname, col.model)
}
\arguments{
\item{file.ext}{Path to the ext file}

\item{return}{The .ext file contains both final parameter
estimates and iterations of the estimates. If
\code{return="pars"} (default) the final estimates are
returned in addition to what other parameter-level information
is found. If \code{return="iterations"}, the iterations are
returned. If \code{return="both"}, both er returned, though in
separate data.frames compiled in a list.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say tibble::as_tibble) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}

\item{modelname}{See ?NMscanData}

\item{col.model}{See ?NMscanData}
}
\value{
A list with a final parameter table and a table of the
    iterations
}
\description{
Read information from Nonmem ext files
}
\keyword{export}
