% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadTab.R
\name{NMreadTab}
\alias{NMreadTab}
\title{Read an output table file from NONMEM}
\usage{
NMreadTab(file, silent = TRUE, tab.count = TRUE, as.fun = NULL, ...)
}
\arguments{
\item{file}{path to NONMEM table file}

\item{silent}{logical stating whether or not information is
printed about what is being done.}

\item{tab.count}{Nonmem includes a counter of tables in the
written data files. These are often not useful. However, if
tab.count is TRUE (default), this will be carried forward and
added as a column called TABLENO.}

\item{as.fun}{The default is to return data in data.tables. Pass a
function in as.fun to convert to something else. If
data.frames are wanted, use as.fun=as.data.frame. See
?runAsFun.}

\item{...}{Arguments passed to fread.}
}
\value{
The Nonmem table data.
}
\description{
Read a table generated by a $TABLE statement in Nonmem. Generally,
these files cannot be read by read.table or similar because
formatting depends on options in the $TABLE statement, and because
Nonmem sometimes includes extra lines in the output that have to
be filtered out. NMreadTab can do this automatically based on the
table file alone.
}
\details{
The actual reading of data is based on
    data.table::fread. Generally, the function is fast thanks to
    data.table.
}
\seealso{
Other DataRead: 
\code{\link{NMreadCsv}()},
\code{\link{NMscanData}()},
\code{\link{NMscanTables}()},
\code{\link{NMtransInput}()}
}
\concept{DataRead}
