% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMscanTables.R
\name{NMscanTables}
\alias{NMscanTables}
\title{read all output data tables in nonmem run}
\usage{
NMscanTables(file, details = F, as.dt = TRUE, quiet = FALSE,
  debug = FALSE)
}
\arguments{
\item{file}{the nonmem file to read (normally .mod or .lst)}

\item{details}{If TRUE, metadata is added to output. In this case, you get a
list. I'd say, enable if doing programming.}

\item{as.dt}{return data.tables? Tables will not be keyed.}

\item{quiet}{The default is to give some information along the way on what
data is found. But consider setting this to TRUE for non-interactive
use.}

\item{debug}{Start by calling browser()?}
}
\value{
A list of all the tables as data.frames. If details=TRUE, this is in
    one element, called data, and meta is another element. If not, only the
    element corresponding to data is returned.
}
\description{
read all output data tables in nonmem run
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadCsv}},
  \code{\link{NMreadTab}}, \code{\link{NMtransFilters}},
  \code{\link{NMtransInput}}, \code{\link{NMundoReplace}},
  \code{\link{NMwriteData}}
}
\concept{Nonmem}
