% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMtransInp.R
\name{NMtransInp}
\alias{NMtransInp}
\title{translate the column names according to the $INPUT section of a control stream}
\usage{
NMtransInp(data, file, translate = TRUE, recover.cols = TRUE, quiet = FALSE)
}
\arguments{
\item{data}{the data to translate}

\item{file}{the list file or control stream}

\item{translate}{logical. Do translation according to Nonmem code
or not? If not, an overview of column names in data and in
Nonmem code is still returned with the data.}

\item{recover.cols}{recover columns that were not used in the
NONMEM control stream? Default is TRUE. Can only be negative
when translate=FALSE.}

\item{quiet}{Suppress warnings about data columns?}
}
\value{
data with column names translated as specified by nonmem
    control stream. Class same as for 'data' argument. Class
    data.table.
}
\description{
translate the column names according to the $INPUT section of a control stream
}
\keyword{internal}
