% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSAME.R
\name{NSAME}
\alias{NSAME}
\title{How many SAME reps?}
\usage{
NSAME(x)
}
\arguments{
\item{x}{A character string to test}
}
\description{
How many SAME reps?
}
\examples{
NMdata:::NSAME("SAME")
NMdata:::NSAME(" SAME ")
NMdata:::NSAME("SAME(1)")
NMdata:::NSAME("SAME (1)")
NMdata:::NSAME("  SAME   (  123 ) ")
NMdata:::NSAME("SAME()")
NMdata:::NSAME("SAME (a)")
NMdata:::NSAME("NOT SAME")
}
\keyword{internal}
