% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objInfo.R
\name{objInfo}
\alias{objInfo}
\title{Get stamped info about object}
\usage{
objInfo(data)
}
\arguments{
\item{data}{The dataset to receive info about}
}
\description{
Get stamped info about object
}
\seealso{
stampObj

Other DataGen: \code{\link{flagsAssign}},
  \code{\link{flagsCount}}, \code{\link{stampObj}}
}
\concept{DataGen}
