% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot raw nonmem data}
\usage{
plot_data(database, time = "TIME", dv = "DV", xmax = "", logy = "", strat = "")
}
\arguments{
\item{database}{name of the database}

\item{time}{Name of the time column you want on the x-axis (e.g. 'TAD')}

\item{dv}{Dependend variable (e.g. concentration 'DV')}

\item{xmax}{optional: limit the x-axis to a value (only upper) (e.g. xmax=24)}

\item{logy}{optional: log10 transform of y scale ('yes', 'y', 'true','T')}

\item{strat}{optional: stratify, create facets (e.g. strat = 'CMT')}
}
\description{
Plot raw nonmem data
}
