% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder2.R
\name{reorder2}
\alias{reorder2}
\title{reorder with respect to multiple vectors}
\usage{
reorder2(x, ...)
}
\arguments{
\item{x}{The vector to be reordered.}

\item{...}{vectors or expressions of vectors to reorder by.}
}
\value{
A factor
}
\description{
Like reorder() but takes multiple ordering arguments. The reordering can be
with respect to expressions.
}
