% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMdata_utils.R
\name{runAsFun}
\alias{runAsFun}
\title{Determine as.fun to use based on argument and options}
\usage{
runAsFun(data, as.fun)
}
\arguments{
\item{data}{Dataset to possibly convert.}

\item{as.fun}{A function to apply if not NULL.}
}
\value{
Possibly converted data.
}
\description{
Determine as.fun to use based on argument and options
}
