% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{tad}
\alias{tad}
\title{Calculate time after dose}
\usage{
tad(date, time)
}
\arguments{
\item{DATE}{Date}

\item{TIME}{Time}

\item{method}{method to calculate TAD}
}
\description{
Calculate time after dose
}
\examples{
-
}
\keyword{TAD}
