% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_figures.R
\name{xp_figures}
\alias{xp_figures}
\title{xpose4 validation figures using xpose4}
\usage{
xp_figures(
  run,
  stratify = "",
  log = "",
  idv = "",
  col = "black",
  pch = 16,
  cex = 1
)
}
\arguments{
\item{stratify}{}
}
\description{
xpose4 validation figures using xpose4
}
