% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_figures.R
\name{xp_figures}
\alias{xp_figures}
\title{xpose4 validation figures using xpose4}
\usage{
xp_figures(
  run,
  stratify = "",
  log = "",
  idv = "",
  col = "black",
  pch = 16,
  cex = 1,
  sep_figs = ""
)
}
\arguments{
\item{run}{The run number /(e.g. like '001/)}

\item{stratify}{Stratify on item}

\item{log}{Log transforms data}

\item{idv}{Independent variable which can be changed}

\item{col}{Color}

\item{pch}{pch shape of markers lattice}

\item{cex}{cex size of markers lattice}

\item{sep_figs}{If true this seperates the figures into a list}
}
\description{
xpose4 validation figures using xpose4
}
